git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@16828 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/sixel.c
@@ -712,7 +712,7 @@ static MagickBooleanType sixel_encode_impl(unsigned char *pixels, size_t width,s
         return (MagickFalse);
     }
     sixel_advance(context, nwrite);
-    nwrite = sprintf((char *)context->buffer + context->pos, "\"1;1;%d;%d", width, height);
+    nwrite = sprintf((char *)context->buffer + context->pos, "\"1;1;%d;%d", (int) width, (int) height);
     if (nwrite <= 0) {
         RelinquishNodesAndMap;
         return (MagickFalse);
@@ -720,7 +720,7 @@ static MagickBooleanType sixel_encode_impl(unsigned char *pixels, size_t width,s
     sixel_advance(context, nwrite);
 
     if (ncolors != 2 || keycolor == -1) {
-        for (n = 0; n < ncolors; n++) {
+        for (n = 0; n < (ssize_t) ncolors; n++) {
             /* DECGCI Graphics Color Introducer  # Pc ; Pu; Px; Py; Pz */
             nwrite = sprintf((char *)context->buffer + context->pos, "#%d;2;%d;%d;%d",
                              n,
@@ -739,36 +739,36 @@ static MagickBooleanType sixel_encode_impl(unsigned char *pixels, size_t width,s
         }
     }
 
-    for (y = i = 0; y < height; y++) {
-        for (x = 0; x < width; x++) {
+    for (y = i = 0; y < (ssize_t) height; y++) {
+        for (x = 0; x < (ssize_t) width; x++) {
             pix = pixels[y * width + x];
-            if (pix >= 0 && pix < ncolors && pix != keycolor) {
+            if (pix >= 0 && pix < (ssize_t) ncolors && pix != keycolor) {
                 map[pix * width + x] |= (1 << i);
             }
         }
 
-        if (++i < 6 && (y + 1) < height) {
+        if (++i < 6 && (y + 1) < (ssize_t) height) {
             continue;
         }
 
-        for (c = 0; c < ncolors; c++) {
-            for (left = 0; left < width; left++) {
+        for (c = 0; c < (ssize_t) ncolors; c++) {
+            for (left = 0; left < (ssize_t) width; left++) {
                 if (*(map + c * width + left) == 0) {
                     continue;
                 }
 
-                for (right = left + 1; right < width; right++) {
+                for (right = left + 1; right < (ssize_t) width; right++) {
                     if (*(map + c * width + right) != 0) {
                         continue;
                     }
 
-                    for (n = 1; (right + n) < width; n++) {
+                    for (n = 1; (right + n) < (ssize_t) width; n++) {
                         if (*(map + c * width + right + n) != 0) {
                             break;
                         }
                     }
 
-                    if (n >= 10 || right + n >= width) {
+                    if (n >= 10 || right + n >= (ssize_t) width) {
                         break;
                     }
                     right = right + n - 1;
@@ -1031,7 +1031,7 @@ static Image *ReadSIXELImage(const ImageInfo *image_info,ExceptionInfo *exceptio
       sixel_palette=(unsigned char *) RelinquishMagickMemory(sixel_palette);
       ThrowReaderException(ResourceLimitError,"MemoryAllocationFailed");
     }
-  for (i = 0; i < image->colors; ++i) {
+  for (i = 0; i < (ssize_t) image->colors; ++i) {
     image->colormap[i].red   = ScaleCharToQuantum(sixel_palette[i * 4 + 0]);
     image->colormap[i].green = ScaleCharToQuantum(sixel_palette[i * 4 + 1]);
     image->colormap[i].blue  = ScaleCharToQuantum(sixel_palette[i * 4 + 2]);
